+++ components/radio/src/auro-radio.js
@@ -40,6 +40,7 @@ export class AuroRadio extends LitElement {
     this.error = false;
     this.onDark = false;
     this.tabIndex = -1;
+    this.touched = false;
 
     /**
      * @private
@@ -85,6 +86,16 @@ export class AuroRadio extends LitElement {
       tabIndex: {
         type: Number,
         reflect: true
+      },
+
+      /**
+       * Whether or not the radio button has been touched by the user.
+       * @private
+       */
+      touched: {
+        type: Boolean,
+        reflect: true,
+        attribute: false
       }
     };
   }
@@ -136,6 +147,7 @@ export class AuroRadio extends LitElement {
    * @returns {void}
    */
   handleFocus(event) {
+    this.touched = true;
     this.dispatchEvent(new CustomEvent('focusSelected', {
       bubbles: true,
       composed: true,
@@ -162,6 +174,7 @@ export class AuroRadio extends LitElement {
    * @returns {void}
    */
   reset() {
+    this.touched = false;
     this.checked = false;
     this.error = false;
   }